Adjusted the display of the old and new outlines. Made corners on the new outlines bigger, fixed the missing display of old outline sides. Adjusted the positions to centralise the outline points on the middle of blocks for all but corners on the old outlines which have precise decimalised locations.  Changed the conversion of MC coordinates to geographical coordinates to sue the centre of blocks as this is now represents what is displayed to the player in game

+++ src/main/java/me/bteuk/progressmapper/Utils.java
@@ -114,7 +114,7 @@ public class Utils {
         return longLat;
     }
 
-    public static double[] convertFromMcCoordinatesToGeometricCoordinates(long x, long z)
+    public static double[] convertFromMcCoordinatesToGeometricCoordinates(double x, double z)
     {
         double[] longLat = null;
